feat(answer): added author field to answers and improved answer management
- Updated AnswerController and AnswerService to handle answers with author data.
- Modified DTOs GetAllAnswerByTopic and GetAnswer for consistent response structure.
- Updated GetTopic to include associated answers with author details.
- Added migration V6 to include author field in answers table.

+++ src/main/java/com/eduforum/api/forum_api/domain/answer/service/AnswerService.java
@@ -7,7 +7,10 @@ import com.eduforum.api.forum_api.domain.answer.dtos.UpdateAnswerDTO;
 import com.eduforum.api.forum_api.domain.answer.model.Answer;
 import com.eduforum.api.forum_api.domain.answer.repository.AnswerRepository;
 import com.eduforum.api.forum_api.domain.serializer.Success;
+import com.eduforum.api.forum_api.domain.topic.model.Topic;
 import com.eduforum.api.forum_api.domain.topic.service.TopicService;
+import com.eduforum.api.forum_api.domain.user.service.UserService;
+import com.eduforum.api.forum_api.infra.errors.ForbiddenException;
 import com.eduforum.api.forum_api.infra.errors.ValidateException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -19,22 +22,26 @@ public class AnswerService {
 
   private final AnswerRepository answerRepository;
   private final TopicService topicService;
+  private final UserService userService;
 
   @Autowired
-  public AnswerService(AnswerRepository answerRepository, TopicService topicService) {
+  public AnswerService(AnswerRepository answerRepository, TopicService topicService, UserService userService) {
     this.answerRepository = answerRepository;
     this.topicService = topicService;
+    this.userService = userService;
   }
 
-  public GetAnswer createAnswer(CreateAnswerDTO payload){
+  public GetAnswer createAnswer(CreateAnswerDTO payload, String email){
+    var user = this.userService.findUserByEmail(email);
+
     var topic = this.topicService.findTopic(payload.idTopic());
-    Answer answer = this.answerRepository.save(new Answer(payload.message(), topic));
+    Answer answer = this.answerRepository.save(new Answer(payload.message(), topic, user));
     return new GetAnswer(answer);
   }
 
-  public Page<GetAllAnswerByTopic> getAllAnswerByTopic(Pageable pageable, Long idTopic) {
+  public Page<GetAnswer> getAllAnswerByTopic(Pageable pageable, Long idTopic) {
     var topic = this.topicService.findTopic(idTopic);
-    return this.answerRepository.findByTopicId(pageable, topic.getId()).map(GetAllAnswerByTopic::new);
+    return this.answerRepository.findByTopicId(pageable, topic.getId()).map(GetAnswer::new);
   }
 
   public Answer findAnswer(Long id) {
@@ -43,15 +50,26 @@ public class AnswerService {
     );
   }
 
-  public GetAnswer updateAnswer(Long id, UpdateAnswerDTO payload) {
+  public GetAnswer updateAnswer(Long id, UpdateAnswerDTO payload, String email) {
     Answer answer = this.findAnswer(id);
+    verifyAuthor(answer, email);
+
     answer.updateAnswer(payload);
     return new GetAnswer(answer);
   }
 
-  public Success deleteAnswer(Long id) {
+  public Success deleteAnswer(Long id, String email) {
     Answer answer = this.findAnswer(id);
+    verifyAuthor(answer, email);
+
     this.answerRepository.delete(answer);
     return new Success(true, "answer with id (" + id + ") deleted");
+  }
+
+  public void verifyAuthor(Answer answer, String email) {
+    var user = this.userService.findUserByEmail(email);
+    if (answer.getUser() != user) {
+      throw new ForbiddenException("permission denied.");
+    }
   }
 }
